+++ src/main/java/antigravity/web/argumentresolver/UserInfo.java
@@ -1,4 +1,4 @@
-package antigravity.web;
+package antigravity.web.argumentresolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
